Moved the Task class inside the communication package in order to prevent dependency cycles
